Install basic system health monitoring tools
Add a section to install sysstat and nmon to enable system health and performance monitoring utilities. Sysstat adds the sar command to the system and nmon is an interactive monitor. Sysstat also requires one change to /etc/default/sysstat:

ENABLED="false"

must be changed to

ENABLED="true"

Without this change, the sar command does not work. The sed command makes the required change in the file to enable system statistics to be collected.