Require users understanding before install.
In an effort to offer further transparency for the users of understanding the underlying requirements in this script (, this changes the curlinstall.sh script to require the user to understand the risks and responsibilities of running the installer. Users must explicitly manifest their acknowledgement that their OS and or any packages with available updates in configured repositories may be updated and or upgraded. Any other input causes abort of the install script.

Signed-off-by: YipYup <p.tanner.w@gmail.com>