rclone tag fix
removed the use of: "--delete-empty-src-dirs" as it would delete the folders inside "/mnt/move" so some programs would have trouble moving stuff, as the folder is trying to reach would no longer be available - instead replaced it with the find command. So parent folders in the move folder is not deleted.