fix: invalid watchtower app list
The "sed" arguments provided only work if the apps are on separate
lines. Which is not how watchtowergen() builds the list. The list
is a single space delimited list of apps "1 2 3". This change also
fixes an issue where multiple calls to watchtowergen() keep
appending the apps to the file (i.e. "1 2 3 1 2 3 1 2 3").